+++ lib/src/compile.dart
@@ -3,6 +3,9 @@
 // https://opensource.org/licenses/MIT.
 
 import 'ast/sass.dart';
+import 'importer.dart';
+import 'importer/filesystem.dart';
+import 'importer/package.dart';
 import 'io.dart';
 import 'sync_package_resolver.dart';
 import 'util/path.dart';
@@ -14,6 +17,7 @@ import 'visitor/serialize.dart';
 CompileResult compile(String path,
         {bool indented,
         bool color: false,
+        Iterable<Importer> importers,
         SyncPackageResolver packageResolver,
         Iterable<String> loadPaths,
         OutputStyle style,
@@ -23,9 +27,11 @@ CompileResult compile(String path,
     compileString(readFile(path),
         indented: indented ?? p.extension(path) == '.sass',
         color: color,
+        importers: importers,
         packageResolver: packageResolver,
-        style: style,
         loadPaths: loadPaths,
+        importer: new FilesystemImporter('.'),
+        style: style,
         useSpaces: useSpaces,
         indentWidth: indentWidth,
         lineFeed: lineFeed,
@@ -36,8 +42,10 @@ CompileResult compile(String path,
 CompileResult compileString(String source,
     {bool indented: false,
     bool color: false,
+    Iterable<Importer> importers,
     SyncPackageResolver packageResolver,
     Iterable<String> loadPaths,
+    Importer importer,
     OutputStyle style,
     bool useSpaces: true,
     int indentWidth,
@@ -46,15 +54,24 @@ CompileResult compileString(String source,
   var sassTree = indented
       ? new Stylesheet.parseSass(source, url: url, color: color)
       : new Stylesheet.parseScss(source, url: url, color: color);
+
+  var importerList = (importers?.toList() ?? []);
+  if (loadPaths != null) {
+    importerList.addAll(loadPaths.map((path) => new FilesystemImporter(path)));
+  }
+  if (packageResolver != null) {
+    importerList.add(new PackageImporter(packageResolver));
+  }
+
   var evaluateResult = evaluate(sassTree,
-      color: color, packageResolver: packageResolver, loadPaths: loadPaths);
+      importers: importerList, importer: importer, color: color);
   var css = serialize(evaluateResult.stylesheet,
       style: style,
       useSpaces: useSpaces,
       indentWidth: indentWidth,
       lineFeed: lineFeed);
 
-  return new CompileResult(css, evaluateResult.includedUrls);
+  return new CompileResult(css, evaluateResult.includedFiles);
 }
 
 /// The result of compiling a Sass document to CSS, along with metadata about
@@ -63,9 +80,12 @@ class CompileResult {
   /// The compiled CSS.
   final String css;
 
-  /// The URLs that were loaded during the compilation, including the main
-  /// file's.
-  final Set<Uri> includedUrls;
+  /// The set that will eventually populate the JS API's
+  /// `result.stats.includedFiles` field.
+  ///
+  /// For filesystem imports, this contains the import path. For all other
+  /// imports, it contains the URL passed to the `@import`.
+  final Set<String> includedFiles;
 
-  CompileResult(this.css, this.includedUrls);
+  CompileResult(this.css, this.includedFiles);
 }